add server.js testing to catalogue
