+++ catalogue/webapp/.dockerignore
@@ -1,0 +1,3 @@
+**
+common/**
+catalogue/webapp/**
