+++ catalogue/webapp/__mocks__/next.js
@@ -1,0 +1,9 @@
+module.exports = () => ({
+  prepare: () => Promise.resolve(),
+  getRequestHandler: () => (req, res) => Promise.resolve(),
+  render: (ctx, next) => {
+    ctx.status = 200;
+    ctx.body = 'ok';
+    return next();
+  }
+});
