+++ catalogue/webapp/package.json
@@ -13,10 +13,12 @@
     "@babel/preset-flow": "^7.0.0",
     "@weco/common": "../../common",
     "autoprefixer": "^8.3.0",
+    "babel-core": "^7.0.0-bridge",
+    "babel-jest": "^23.6.0",
     "cookies": "^0.7.1",
     "css-loader": "^0.28.11",
     "isomorphic-unfetch": "^2.0.0",
-    "jest": "^23.2.0",
+    "jest": "^23.6.0",
     "koa": "^2.5.0",
     "koa-router": "^7.4.0",
     "next": "^7.0.2",
@@ -28,6 +30,8 @@
     "react-dom": "^16.4.1",
     "react-test-renderer": "^16.4.1",
     "react-transition-group": "^2.4.0",
-    "sass-loader": "^7.0.1"
+    "sass-loader": "^7.0.1",
+    "supertest": "^3.3.0",
+    "webpack": "^4.25.1"
   }
 }
