--- catalogue/webapp/test/WorkPage.test.js
@@ -1,13 +1,0 @@
-// @flow
-import TestRenderer from 'react-test-renderer';
-import {getWork} from '../services/catalogue/works';
-import {WorkPage} from '../pages/work';
-
-it('renders <WorkPage /> with an catalogue API work response', async () => {
-  const work = await getWork({id: 'a3gvp92j'});
-  // TODO: Work with Next.js dynamic(import(...))
-  const WorkPageComponent = TestRenderer.create(
-    <WorkPage work={work} page={null} previousQueryString={null} version={1} />
-  );
-  expect(WorkPageComponent).not.toBe(null);
-});
